requirements.txt
pandas==1.3.3
streamlit==1.5.0
plotly==5.3.1
pillow==9.0.0
openpyxl>=3.1.0
